Use authenticated repositories and packages
BugLink: https://launchpad.net/bugs/1836041

Ensure that all of the additionally configured repositories and
installed packages needed to construct a kernel snap are authenticated
by apt.

The Makefile improperly used the --allow-insecure-repositories and
--allow-unauthenticated apt options when setting up the build chroot. An
attacker with control over the network between the build machine and the
Ubuntu archive or the snappy-dev/image PPA could use this to perform a
man-in-the-middle attack to install malicious packages in the build
chroot.

Such an attack is unlikely for the official Ubuntu kernel snap builds
since the Launchpad buildd infrastructure and the network communication
with the Ubuntu archive and Launchpad PPAs is tightly controlled.
However, end-users may use this Makefile to build their own kernel snaps
and have no guarantees about the communication with the archive or PPAs.

Store a copy of the snappy-dev/image PPA's public signing key alongside
the Makefile so that the public signing key can be added to apt as part
of the build process. Finally, remove all uses of
--allow-insecure-repositories and --allow-unauthenticated when invoking
apt commands.

CVE-2019-11480

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>