Find the meta package in pocket using the abi as a key
If the version we are trying to build is the version in -updates but the
version is higher in -proposed we will install the wrong version and bail
out.  Rather find the appropriate meta package version which will install
the kernel we are trying to package and install that explicitly by version.

Signed-off-by: Andy Whitcroft <apw@canonical.com>