Restructure version-check prerequisites
Right now we are in the unfortunate situation of trying to build
using both the "vmlinuz" and "efi" kernel image formats from a
single snap build recipe, depending on the architecture. The fix
for this is likely to split these into different recipes, but as
a short-term fix, restructure the way the version-check
prerequisites are define to support the current state of affairs.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>