Split Makefile into per-image-format files
Rather than cluttering the makefile with conditionals, split it
up into two makefiles, one for the "efi" image type and one for
the "vmlinuz" type. Keep common definitions in the top-level
makefile, and include the appropriate format-specific makefile
based on the value of KERNEL_IMAGE_FORMAT.

Note that we are left with no generic way to determine the kernel
flavour, not that we really had a good, generic way to do it
before. Recipes are now epected to specify FLAVOUR in the
make-parameters section of snapcraft.yaml.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>