Move common install code to shared makefile
Now that use of $(KVER) has been eliminated, only the
installation of the kernel image and initrd differ between the
install targets of the two makefiles. Move the common code to a
common install target, and make this depend on an install-image
target which is defined in the format-specific makefiles.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>