Cleanup arch-specific handling
BugLink: https://bugs.launchpad.net/bugs/1907056

Currently, we do arch-specific things for pi-kernel but that's legacy
cruft that is no longer necessary. Also, the soon-to-be-released snapd
and gadget will use the DTBs (and overlays) from the pi-kernel snap so
what we really need is generic DTB handling. Let's do that.

Signed-off-by: Juerg Haefliger <juergh@canonical.com>