Remove unnecessary firmware binaries
BugLink: https://bugs.launchpad.net/bugs/1951422
BugLink: https://bugs.launchpad.net/bugs/1951423

linux-firmware provides a lot of binaries that are not needed/required and
just blow up the size of the kernel snap. In order to reduce the snap size,
remove all firmware binaries that are not referenced by the kernel. But
only if TRIM_FIRMWARE is set to 'true'.

Note that older kernels don't reference brcm firmware config files so
copy the whole directory content to be on the safe side.

Signed-off-by: Juerg Haefliger <juergh@canonical.com>
Acked-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Juerg Haefliger <juergh@canonical.com>